Sun, Jul 21, 2024, 8:20 PM +05:30

+++ rough/rough.c
@@ -9,5 +9,5 @@
 
 int main(void)
 {
-    int i = 55.5;
+    int i = (int)55.5;
 }
